♻️ Refactor: 코드리뷰 대응 완료
