+++ src/app/(main)/activity/[id]/components/activity-info-header/index.tsx
@@ -1,7 +1,5 @@
-'use client';
-
-import { useEffect, useRef, useState } from 'react';
-import { IcMap, IcMore, IcStar } from '@/shared/assets/icons';
+import KebabDropdown from '@/app/(main)/activity/[id]/components/activity-info-header/KebabDropdown';
+import { IcMap, IcStar } from '@/shared/assets/icons';
 import { cn } from '@/shared/utils/cn';
 
 export interface ActivityInfoHeaderProps {
@@ -15,78 +13,6 @@ export interface ActivityInfoHeaderProps {
   onEdit?: () => void;
   onDelete?: () => void;
   className?: string;
-}
-
-interface KebabDropdownProps {
-  onEdit?: () => void;
-  onDelete?: () => void;
-}
-
-function KebabDropdown({ onEdit, onDelete }: KebabDropdownProps) {
-  const [isOpen, setIsOpen] = useState(false);
-  const containerRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    function handleClickOutside(e: MouseEvent) {
-      if (
-        containerRef.current &&
-        !containerRef.current.contains(e.target as Node)
-      ) {
-        setIsOpen(false);
-      }
-    }
-    if (isOpen) {
-      document.addEventListener('mousedown', handleClickOutside);
-    }
-    return () => document.removeEventListener('mousedown', handleClickOutside);
-  }, [isOpen]);
-
-  return (
-    <div ref={containerRef} className="relative">
-      <button
-        type="button"
-        onClick={() => setIsOpen((v) => !v)}
-        className="flex h-7 w-7 cursor-pointer items-center justify-center rounded-md text-gray-950 transition-colors hover:bg-gray-50"
-        aria-label="더보기"
-        aria-expanded={isOpen}
-        aria-haspopup="menu"
-      >
-        <IcMore aria-hidden="true" className="size-7 shrink-0 text-gray-950" />
-      </button>
-
-      {isOpen && (
-        <div
-          role="menu"
-          className="absolute top-full right-0 z-50 mt-1 h-[73px] w-[63px] overflow-hidden rounded-lg border border-gray-100 bg-white md:h-[108px] md:w-24"
-        >
-          <div className="grid h-full grid-rows-2 divide-y divide-gray-100">
-            <button
-              type="button"
-              role="menuitem"
-              onClick={() => {
-                setIsOpen(false);
-                onEdit?.();
-              }}
-              className="typo-md-medium hover:bg-gray-25 md:typo-lg-medium flex h-full w-full cursor-pointer items-center justify-center text-center text-gray-950 transition-colors"
-            >
-              수정하기
-            </button>
-            <button
-              type="button"
-              role="menuitem"
-              onClick={() => {
-                setIsOpen(false);
-                onDelete?.();
-              }}
-              className="typo-md-medium hover:bg-gray-25 md:typo-lg-medium flex h-full w-full cursor-pointer items-center justify-center text-center text-gray-950 transition-colors"
-            >
-              삭제하기
-            </button>
-          </div>
-        </div>
-      )}
-    </div>
-  );
 }
 
 /**
@@ -111,18 +37,20 @@ export function ActivityInfoHeader({
   onDelete,
   className,
 }: ActivityInfoHeaderProps) {
+  const formattedRating = Number.isFinite(rating) ? rating.toFixed(1) : '-';
+
   return (
     <div className={cn('flex flex-col', className)}>
       {/* 카테고리 + 케밥 */}
       <div className="flex items-center justify-between">
-        <span className="typo-sm-medium md:typo-md-medium tracking-[-0.025em] text-gray-700">
+        <span className="typo-sm-medium md:typo-md-medium tracking-tight text-gray-700">
           {category}
         </span>
         {isOwner && <KebabDropdown onEdit={onEdit} onDelete={onDelete} />}
       </div>
 
       {/* 제목 */}
-      <h2 className="typo-2lg-bold md:typo-2xl-bold mt-1 tracking-[-0.025em] text-gray-950">
+      <h2 className="typo-2lg-bold md:typo-2xl-bold mt-1 tracking-tight text-gray-950">
         {title}
       </h2>
 
@@ -132,10 +60,10 @@ export function ActivityInfoHeader({
           aria-hidden="true"
           className="size-4 shrink-0 text-yellow-500"
         />
-        <span className="typo-md-medium tracking-[-0.025em] text-gray-700">
-          {rating}
+        <span className="typo-md-medium tracking-tight text-gray-700">
+          {formattedRating}
         </span>
-        <span className="typo-md-medium tracking-[-0.025em] text-gray-700">
+        <span className="typo-md-medium tracking-tight text-gray-700">
           ({reviewCount})
         </span>
       </div>
@@ -143,7 +71,7 @@ export function ActivityInfoHeader({
       {/* 위치 */}
       <div className="mt-0.5 flex items-center gap-1">
         <IcMap aria-hidden="true" className="size-4 shrink-0 text-black" />
-        <span className="typo-md-medium tracking-[-0.025em] text-gray-700">
+        <span className="typo-md-medium tracking-tight text-gray-700">
           {address}
         </span>
       </div>
